refactored modelContainer.py -> model_container,py
