+++ batchglm/train/numpy/base_glm/model_container.py
