+++ batchglm/train/numpy/glm_nb/__init__.py
@@ -1,2 +1,2 @@
 from .estimator import Estimator
-from .modelContainer import ModelContainer
+from .model_container import ModelContainer
