+++ batchglm/train/numpy/glm_nb/model_container.py
